refactor: move NiftiAbnormalVoxelData to voxel_data_adapters.py
- Relocate `NiftiAbnormalVoxelData` class from `staple_segmenter.py` to `voxel_data_adapters.py`.
- Simplify imports and update references across modules.
- Adjust and clean up unit tests to reflect the class relocation and enhancements.
