+++ oxytcmri/interface/mri/staple_segmenter.py
@@ -7,158 +7,13 @@ import logging
 import subprocess
 import tempfile
 from pathlib import Path
-from typing import List, cast, Optional, Tuple, Dict
+from typing import List
 
-import numpy
-
-from oxytcmri.domain.entities.mri import VoxelData
 from oxytcmri.domain.use_cases.segment_dti_abnormal_values import SegmentationMerger
 from oxytcmri.domain.entities.dti_lesions import DTIAbnormalValues, AbnormalVoxelData, AbnormalValueType
-from oxytcmri.interface.mri.voxel_data_adapters import NiftiVoxelData
+from oxytcmri.interface.mri.voxel_data_adapters import NiftiAbnormalVoxelData
 
 logger = logging.getLogger(__name__)
-
-
-class NiftiAbnormalVoxelData(AbnormalVoxelData, NiftiVoxelData[int]):
-    """
-    NiftiAbnormalVoxelData combines AbnormalVoxelData and NiftiVoxelData functionalities.
-    This class allows storing, manipulating, and persisting abnormal data in NIfTI files.
-    """
-
-    def __init__(self,
-                 source_voxel_data: VoxelData[float],
-                 nifti_path: Path,
-                 abnormal_voxels: Optional[Dict[Tuple[int, int, int], AbnormalValueType]] = None) -> None:
-        """
-        Initialize a NiftiAbnormalVoxelData instance.
-        
-        Parameters
-        ----------
-        source_voxel_data : VoxelData[float]
-            Source voxel data
-        nifti_path : Path
-            Path to the NIfTI file
-        abnormal_voxels : Optional[Dict[Tuple[int, int, int], AbnormalValueType]]
-            Dictionary of abnormal voxels (optional)
-        """
-        # Initialize both parent classes
-        AbnormalVoxelData.__init__(self, source_voxel_data)
-        NiftiVoxelData.__init__(self, nifti_path)
-
-        # If an abnormal voxels dictionary is provided, use it
-        if abnormal_voxels is not None:
-            self.abnormal_voxels = abnormal_voxels
-
-        # Otherwise, load data from the NIfTI file
-        elif nifti_path.exists():
-            self._load_abnormal_data_from_nifti()
-
-    def _load_abnormal_data_from_nifti(self) -> None:
-        """
-        Load abnormal data from the NIfTI file.
-        Converts integer values to AbnormalValueType.
-        """
-        # Get the dimensions of the NIfTI file
-        dimensions = self.get_dimensions()
-
-        # Iterate through all values and convert them to AbnormalValueType
-        for x in range(dimensions[0]):
-            for y in range(dimensions[1]):
-                for z in range(dimensions[2]):
-                    value = NiftiVoxelData.get_value_at(self, x, y, z)
-                    # Check if the value is close to an integer
-                    try:
-                        rounded_value = numpy.round(value)
-                    except TypeError as e:
-                        raise ValueError(f"Invalid value in voxel data {self} "
-                                         f"at (x,y,z) = ({x}, {y}, {z}): {value}") from e
-                    if abs(rounded_value - value) > 0.25:
-                        logger.warning(
-                            f"Rounding value at (x,y,z) = ({x}, {y}, {z}) "
-                            f"from {value} to {rounded_value}."
-                        )
-                    # Convert the integer to AbnormalValueType
-                    try:
-                        abnormal_type = AbnormalValueType.from_integer(int(rounded_value))
-                        if abnormal_type is not None:
-                            self.abnormal_voxels[(x, y, z)] = abnormal_type
-                    except ValueError:
-                        # Ignore values that cannot be converted
-                        pass
-
-    @classmethod
-    def from_abnormal_voxel_data(cls,
-                                 abnormal_voxel_data: AbnormalVoxelData,
-                                 nifti_path: Path) -> "NiftiAbnormalVoxelData":
-        """
-        Create a NiftiAbnormalVoxelData from an existing AbnormalVoxelData.
-        
-        Parameters
-        ----------
-        abnormal_voxel_data : AbnormalVoxelData
-            Abnormal data to convert
-        nifti_path : Path
-            Path where to save the NIfTI file
-            
-        Returns
-        -------
-        NiftiAbnormalVoxelData
-            Instance with the converted abnormal data
-        """
-        # Convert the AbnormalVoxelData to a numpy array of integers
-        integer_data = cls._convert_to_integer_numpy_array(abnormal_voxel_data)
-
-        source_voxel_data = abnormal_voxel_data.get_source_voxel_data()
-        if not isinstance(source_voxel_data, NiftiVoxelData):
-            raise ValueError("Source voxel data must be of type NiftiVoxelData")
-
-        # Create a NIfTI file with the metadata from the source file
-        NiftiVoxelData.create_with_same_metadata(
-            source_nifti=cast(NiftiVoxelData, source_voxel_data),
-            output_path=nifti_path,
-            data=integer_data,
-        )
-
-        # Create and return a new NiftiAbnormalVoxelData
-        return cls(
-            source_voxel_data=source_voxel_data,
-            nifti_path=nifti_path,
-            abnormal_voxels=abnormal_voxel_data.abnormal_voxels.copy()
-        )
-
-    @staticmethod
-    def _convert_to_integer_numpy_array(voxel_data: AbnormalVoxelData) -> numpy.ndarray:
-        """
-        Convert an AbnormalVoxelData to a numpy array of integers.
-        
-        Parameters
-        ----------
-        voxel_data : AbnormalVoxelData
-            Abnormal data to convert
-            
-        Returns
-        -------
-        numpy.ndarray
-            Numpy array of integers representing the abnormalities
-        """
-        data_dimensions = voxel_data.get_dimensions()
-
-        # Create a numpy array of the same shape as the original data
-        data = numpy.zeros(data_dimensions, dtype=numpy.int32)
-
-        # Set the values based on the AbnormalValueType
-        for x in range(data_dimensions[0]):
-            for y in range(data_dimensions[1]):
-                for z in range(data_dimensions[2]):
-                    value = voxel_data.get_value_at(x, y, z)
-                    if value is not None:
-                        try:
-                            data[x, y, z] = value.to_integer()
-                        except ValueError as e:
-                            raise ValueError(f"Invalid value in voxel data {voxel_data} "
-                                             f"at (x,y,z) = ({x}, {y}, {z}): {value}") from e
-
-        return data
 
 
 class TemporaryFilesHandler:
