+++ oxytcmri/tests/unit/infrastructure/gateways/test_sqlmodel_data_gateway.py
@@ -8,8 +8,7 @@ from oxytcmri.domain.entities.center import Center
 from oxytcmri.domain.entities.dti_lesions import AbnormalVoxelData, DTIAbnormalValues
 from oxytcmri.domain.entities.mri import DTIMap, DTIMetric, MRIExamId
 from oxytcmri.infrastructure.gateways.sqlmodel_data_gateway import SQLModelSQLiteDataGateway
-from oxytcmri.interface.mri.staple_segmenter import NiftiAbnormalVoxelData
-from oxytcmri.interface.mri.voxel_data_adapters import NiftiVoxelData
+from oxytcmri.interface.mri.voxel_data_adapters import NiftiVoxelData, NiftiAbnormalVoxelData
 from oxytcmri.tests.fixtures import path_to_test_data_folder
