+++ oxytcmri/tests/unit/interface/mri/test_unit_staple_segmenter.py
@@ -8,8 +8,8 @@ import pytest
 
 from oxytcmri.domain.entities.dti_lesions import DTIAbnormalValues, AbnormalValueType
 from oxytcmri.domain.entities.mri import DTIMap, DTIMetric, MRIExamId
-from oxytcmri.interface.mri.staple_segmenter import C3DSTAPLESegmentationMerger, NiftiAbnormalVoxelData
-from oxytcmri.interface.mri.voxel_data_adapters import NiftiVoxelData
+from oxytcmri.interface.mri.staple_segmenter import C3DSTAPLESegmentationMerger
+from oxytcmri.interface.mri.voxel_data_adapters import NiftiVoxelData, NiftiAbnormalVoxelData
 from oxytcmri.tests.fixtures import path_to_test_data_folder
